requirements.txt
Flask==3.0.0
openai==1.12.0
stability-sdk==0.8.5
Pillow==10.2.0
telethon==1.34.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.3
flask-cors==4.0.0